refactor(deployment): fold the patch pipeline into the deployment writer
There was no strong reason for the separation, and the cost of it was four private helpers
that were near-copies of the writer's: `#rejectInvalidSdl` byte-identical, `#resolve` a
narrower `#resolveSdl`, `#commitOnChain` the same base64-versus-`deployment.hash` compare as
`ensureDeploymentIsUpToDate` with the branch inverted, and `#sendManifestToProviders` the
same loop with the wallet id spelled differently. Merging is what deletes them.

`patchByUserIdAndDseq` and the five helpers that are genuinely its own — finding the stored
definition, parsing it, serializing it, merging and pruning the secret set, and refusing an
unreferenced supplied name — move onto `DeploymentWriterService`. The four duplicates are
gone and the writer's own helpers do the work; `#resolveSdl` already took both `secrets` and
`isTrialing`, so nothing needed widening. Net −62 production lines.

The ability is passed in as a parameter rather than injected. This service is a
`@singleton()`, so holding a request-scoped auth reference on it would pin one request's
ability for the process lifetime; `544f183f9` moved the repo deliberately away from that, and
the controller already reads `currentUser.id` and passes it down, so the ability follows the
same path.

Nothing in create or the deprecated update changed beyond now having two more collaborators
injected. The deprecated update is scheduled for removal and nothing new is built on it.

**Log context change:** these events now carry `context: "DeploymentWriterService"` instead
of `"DeploymentPatchService"`. Every `event` name is unchanged —
`DEPLOYMENT_PATCH_APPLIED`, `DEPLOYMENT_STORED_SDL_UNPARSEABLE`,
`DEPLOYMENT_PATCHED_SDL_TOO_LARGE` — so anything filtering on those is unaffected, but a
filter on the context needs updating.

All 43 patch tests move into the writer's spec as a nested suite with its own setup, rather
than being folded into the writer's existing one, so none of the 94 tests already there
shifted behaviour. 137 pass.
